+++ test/translator/translator.translate.plural.spec.js
@@ -28,6 +28,24 @@ describe('Translator', () => {
             test_other: 'tests_ja',
           },
         },
+        ar: {
+          translation: {
+            test: 'test_ar',
+            test_few: 'tests_ar_few',
+            test_many: 'tests_ar_many',
+            test_one: 'tests_ar_one',
+            test_two: 'tests_ar_two',
+            test_zero: 'tests_ar_zero',
+            test_other: 'tests_ar_other',
+          },
+        },
+        it: {
+          translation: {
+            test: 'test_it',
+            test_other: 'tests_it_other',
+            test_many: 'tests_it_many', // ordinal
+          },
+        },
       });
       const lu = new LanguageUtils({ fallbackLng: 'en' });
       t = new Translator(
@@ -64,6 +82,52 @@ describe('Translator', () => {
       { args: ['translation:test', { count: 1, lng: 'ja' }], expected: 'tests_ja' },
       { args: ['translation:test', { count: 2, lng: 'ja' }], expected: 'tests_ja' },
       { args: ['translation:test', { count: 10, lng: 'ja' }], expected: 'tests_ja' },
+      { args: ['translation:test', { count: 0, lng: 'ar' }], expected: 'tests_ar_zero' },
+      { args: ['translation:test', { count: 1, lng: 'ar' }], expected: 'tests_ar_one' },
+      { args: ['translation:test', { count: 2, lng: 'ar' }], expected: 'tests_ar_two' },
+      { args: ['translation:test', { count: 3, lng: 'ar' }], expected: 'tests_ar_few' },
+      { args: ['translation:test', { count: 15, lng: 'ar' }], expected: 'tests_ar_many' },
+      { args: ['translation:test', { count: 101, lng: 'ar' }], expected: 'tests_ar_other' },
+      {
+        args: ['translation:test', { count: 0, lng: 'ar', ordinal: true }],
+        expected: 'tests_ar_other',
+      },
+      {
+        args: ['translation:test', { count: 1, lng: 'ar', ordinal: true }],
+        expected: 'tests_ar_other',
+      },
+      {
+        args: ['translation:test', { count: 2, lng: 'ar', ordinal: true }],
+        expected: 'tests_ar_other',
+      },
+      {
+        args: ['translation:test', { count: 3, lng: 'ar', ordinal: true }],
+        expected: 'tests_ar_other',
+      },
+      {
+        args: ['translation:test', { count: 15, lng: 'ar', ordinal: true }],
+        expected: 'tests_ar_other',
+      },
+      { args: ['translation:test', { count: 0, lng: 'it' }], expected: 'tests_it_other' },
+      { args: ['translation:test', { count: 1, lng: 'it' }], expected: 'test_it' },
+      { args: ['translation:test', { count: 2, lng: 'it' }], expected: 'tests_it_other' },
+      { args: ['translation:test', { count: 11, lng: 'it' }], expected: 'tests_it_other' },
+      {
+        args: ['translation:test', { count: 0, lng: 'it', ordinal: true }],
+        expected: 'tests_it_other',
+      },
+      {
+        args: ['translation:test', { count: 1, lng: 'it', ordinal: true }],
+        expected: 'tests_it_other',
+      },
+      {
+        args: ['translation:test', { count: 2, lng: 'it', ordinal: true }],
+        expected: 'tests_it_other',
+      },
+      {
+        args: ['translation:test', { count: 11, lng: 'it', ordinal: true }],
+        expected: 'tests_it_many',
+      },
     ];
 
     tests.forEach((test) => {