+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 **This is a major breaking release:**
 
-- introduce new v4 json format (intl plural forms), simplifyPluralSuffix not used anymore
+- introduce new v4 json format (intl plural forms with ordinal support), simplifyPluralSuffix not used anymore
 - `skipOnVariables` by default now is true
 - automatically detect natural language keys (no need to set nsSeparator or keySeparator to false)
 - remove deprecated whitelist features